Add conditions to run task less often
Previously the XamarinBuildDownloadPartialZip item was getting added if the .aar wasn’t in the Android SDK location.  If the .aar exists already in the download cache, we were still running the task to look for this item.  Now the task would figure out quickly the item didn’t need downloading and was already cached, but with this changed condition we don’t even have to go that far to figure out it’s not needed.