Merge .targets files in build script
We can only have one .targets file in a nupkg that gets auto-imported into the project (it must be named nuget-package-id.targets), and since some of our packages already were using this file, we need to merge the generated targets for Xamarin.Build.Download with these files.  This means the custom build tasks must be built before the nuget-setup task runs.  This changes the build script to build the existing custom tasks, copy their .targets files, and then detect them when generated the .targets files for Xamarin.Build.Download, and merge them together.