Fix —no-version-vectors agai
There were some initial optimizations done to the adding of —no-version-vectors to AndroidResgenExtraArgs, but they were made to the wrong file.  The build script caused the modified file to be overwritten during the build making the changes ineffective.

This commit cleans up where the .targets file lives so the build tasks project reads it from one source, and the build script no longer copies the file around.