Make everything depened on Annotations
This is the new unifying package which all others depend on in support so that there are no javac errors during app compile due to missing referenced annotations.

This also acts as a version locking unification package for all of support libs, so it should no longer be possible to get versions out of sync (eg: previously you could install a different version of CardView).