Rework build script to use .aar’s
Now we are using .aar files directly from the new maven repo.  This simplifies the process as Xamarin.Build.Download is not involved anymore.

There’s still some things we need to manually handle, one of which is including the proguard.txt files directly in the nuget package and specifying a .targets file to include the `@(ProguardConfiguration)` items for them.

We also fix up the Android Manifest files with the same fixes Xamarin.Build.Download used to apply.