Added `AndroidFragmentType` msbuild property
The upcoming code behind generation feature requires knowing the type of `Fragment` to use in some cases.  By default it uses `Android.App.Fragment`, but if the user installs the support fragment package, we want to default to use `Android.Support.V4.App.Fragment` instead.

This inclusion of a msbuild property should override the default and use the support library version instead.