Initial commit using binding project generation tool
This commit marks a different approach to maintaining the bindings, and now uses a binding generation tool currently called AndroidBinderator.

This tool will fetch all the required maven artifact information including dependency chains, binaries, etc, and generate Xamarin.Android binding projects for the maven artifacts based on Razor templates for both the .csproj and .targets files.

This also uses new SDK style projects which makes packaging nugets easier, especially for including the correct dependency chains in the nupkg’s (which is automated from the maven repository info).

This is a work in progress.