[msbuild] don't add @(AndroidJavaLibrary) in class libraries
Context: https://github.com/xamarin/xamarin-android/pull/5477

I found that adding the AndroidX packages to a class library adds some
AndroidX `.jar` files to the build output...

Such as:

https://github.com/xamarin/xamarin-android/blob/597e7169b069eff781d8f29ddb7746b719d7f781/src/Xamarin.Android.Build.Tasks/Xamarin.Android.Legacy.targets#L452-L466

I think that we only want to add `.jar` files to Android application
projects. Otherwise, this is attempting to add these files to the
`__AndroidLibraryProjects__.zip` file that is an `EmbeddedResource` of
Xamarin.Android class libraries.

Let's add a `Condition=" '$(AndroidApplication)' == 'true' "`, so that
these `@(AndroidJavaLibrary)` or `@(AndroidAarLibrary)` items are only
added for Xamarin.Android application projects and not class
libraries.