[msbuild] allow @(AndroidAarLibrary) in class libraries
Fixes: https://github.com/xamarin/AndroidX/issues/267

A minimal combination of these packages causes a build failure in a
class library:

    <PackageReference Include="Xamarin.AndroidX.AppCompat" Version="1.2.0.6" />
    <PackageReference Include="Xamarin.AndroidX.Core" Version="1.3.2.2" />

Fails with:

    (_UpdateAndroidResgenAapt2 target) ->
        ClassLibrary1\obj\Debug\monoandroid10.0\lp\42\jl\res\color\abc_btn_colored_borderless_text_material.xml(21): error APT2260: attribute alpha (aka ClassLibrary1.ClassLibrary1:alpha) not found.
        ...
        ClassLibrary1\obj\Debug\monoandroid10.0\lp\42\jl\res\color\abc_tint_switch_track.xml(22): error APT2260: attribute alpha (aka ClassLibrary1.ClassLibrary1:alpha) not found.
        C:\Program Files (x86)\Microsoft Visual Studio\2019\Enterprise\MSBuild\Xamarin\Android\Xamarin.Android.Aapt2.targets(164,3): error APT2061: failed linking file resources.

This broke in 8625c58a / #239 when I was trying to fix a situation
where we were adding `@(AndroidJavaLibary)` to class libraries. This
was bad because it was *repackaging* the `.jar` from a NuGet package
into the user's library.

The fix in 8625c58a, was to to not add these to class library
projects. We mistakenly thought we should do this for
`@(AndroidAarLibrary)` as well, but that was a bad idea! We need the
`@(AndroidAarLibrary)` for `Resource.designer.cs` generation.

To test for this problem in the future, I added a
`BuildAll/ClassLibrary.csproj` that reproduced the issue. We now build
a class library that references all the AndroidX packages.

I also updated the `NuGet.config` to use a local `packages` folder. I
was seeing my global NuGet cache `%userprofile%\.nuget\packages` being
used when building these samples.