Fix update script to handle misnamed guava Java versions. (#880)
We have missed several new versions of `com.google.guava.guava` because they name their release packages like `33.1.0-android`.  Traditionally, adding any sort of label after a hyphen indicates that the package is a pre-release version so we ignore it.  Thus we have been ignoring their new versions.

Fix our update script to allow a pre-release tag of `-android` to count as a stable release version so we will process `guava` updates.

Additionally, go ahead and bump `guava` to the latest version available. Note we remove the `-android` from the NuGet version, as it is not a pre-release package.