[NuGet.config] Remove wildcard from nuget.org. (#891)
In order to ensure we are testing the packages built during CI/local build we need to remove the wildcard from `nuget.org`.

With the `nuget.org` wildcard it means that anything that cannot be found locally will be pulled from published packages.  We will never know if we are accidentally pulling a published package instead of the intended local one.