Add extended CI test job. (#892)
Our current `BuildAllDotNet` and `BuidAllMauiApp` tests provide a smoke test that we are able to compile an application that consumes *all* of the packages from this repository *simultaneously*.  However, this doesn't surface dependency errors that a user would see if they add *individual* packages to their application, which is the way packages are actually consumed.

This PR provides extended tests that do the following for each individual package:

- Create a new `android`/`maui` template app
- Add one package to the app
- Compile the app

Even though we parallelize the tests across many test agents, this involves compiling over 600 applications, so the tests take a long time to run and cannot be run on every commit.  For now, we will start with this being a manually run CI test that can be kicked off in the CI UI by setting the `Run Extended Tests?` parameter to `true`.

Note that there are currently packages that fail this test.  Our goal will be to fix these in future changes.