Ignore `nuget-diff` error for now. (#917)
Context: https://github.com/xamarin/AndroidX/issues/916

`api-diff` currently fails for `Xamarin.AndroidX.Car.App.App` `1.4.0` which is preventing us from updating it.  The nature of this crash is documented in https://github.com/xamarin/AndroidX/issues/916.

We have chosen not to spend the resources at this time to fix this tooling error.  Instead, we are going to go ahead and update `Xamarin.AndroidX.Car.App.App` to version `1.4.0` and are going to exclude it from running `api-diff`.