﻿Move functionality from `update-config.csx` to `binderator`. (#927)
As the scope of `update-config.csx` has grown, it has become more and more painful that it isn't in an actual project editable in an IDE, making it harder to update or debug.  As it shares code with `binderator`, like reading and writing `config.json`, move the functionality into `binderator`.

To help facilitate running this functionality, add new targets to `cake` for the various operations:

*   `update-config`

    Updates config.json to the latest versions found in Maven.
    
*   `bump-config`

    Increments the NuGet patch version of all packages in config.json.
    
*   `sort-config`

    Sorts config.json file using the canonical sort.
    
*   `published-config`

    Shows which NuGet package versions in config.json have been published to NuGet.org.

Note these run the in-tree version of `binderator`, there is no need to install the .NET Global tool version to run these targets.