Streamline "LICENSE" handling. (#931)
Reduce maintenance costs by creating a generic `PackageLicense.md` file that can be used by all packages.  Instead of listing package-specific licensing, it points to the existing `THIRD-PARTY-NOTICES.txt` file which already contains package-specific licensing.

Additionally, this is done as `PackageLicense.md` instead of the root `LICENSE.md` because it is specific to the packages we ship and not necessarily to the code in this repository.  That is, the package-specific wording does not apply to things like the `binderator` code that exists in this repository.

Also, mentions of `Xamarin` were replaced with `Microsoft`.

Finally, pull in the standard `dotnet` `LICENSE.txt` for the repository level license, a la:
- https://github.com/dotnet/runtime/blob/main/LICENSE.TXT
- https://github.com/dotnet/android/blob/main/LICENSE.TXT