[ci] Make "Sign Archive" a separate stage. (#965)
Today, when preparing a publish release, the "Sign Archive" step always fails the first run, as it relies on the output from the "Build Windows" job, but does not specify the dependency, so it runs before the packages have been built.

Specifying the dependency is better, but it makes the "Build Windows" stage take a lot longer and the test stage(s) cannot run until it completes.

This PR moves the signing to its own stage with the proper dependency set, so it will not run until the packages are ready.  By moving it to its own stage it can also run in parallel with the test stage(s), reducing the overall CI time.