[ci] Create CI option to test running repository on next version of .NET. (#1013)
We waited too late to ensure that this repository builds successfully on .NET 9, causing last-minute fixes in `generator`.  This is something we need to be doing throughout the entire preview cycle, even though we do not ship packages targeting the preview framework.

This PR creates a CI option that can be requested which will run the repository on the preview version of .NET N+1.

Once this test succeeds (hopefully with .NET 9 RTM!) we can enable it as a nightly test to ensure we are continuously testing our .NET previews instead of waiting until the end of the preview cycle.