Suppress binding warnings. (#1025)
Suppress the warnings emitted by `generator` by default.  These 19K+ warnings are clearly not things that we intend to take action on, and hide warnings we ***might*** want to take action on.

This is controlled via the `$(_AndroidIgnoreGeneratorWarnings)` property.

If seeing these warnings is desired on a local build, either:
- Change `$(_AndroidIgnoreGeneratorWarnings)` to `false` in the `Directory.Build.props` file.
- Pass it as a build property:
  - `dotnet build -p:_AndroidIgnoreGeneratorWarnings=false`

Also fix some `IsTrimmable` warnings by not enabling it on `netstandard2.0` projects.