[ci] `macos12` is no longer available, use `macos14`. (#1034)
The Mac pool we use `internal-macos12` is no longer available.  Ideally we could just use the public Azure Pipelines pool, but those machines are underpowered and doubles the (already long!) time taken to build our packages.  Switch to the internal `VSEng-VSMac-Xamarin-Shared` mac pool already in use by `dotnet/android` and MAUI.

`JavaToolInstaller@0` does not seem to work on this image, even though there are many versions of Java SDK already installed on the image.  Set `JAVA_HOME` to Java SDK 11 so we can build.

Additionally, the `tools-executive-oreder-csv-and-markdown` Cake task does not work on these images because it expects `gradle` and other tools to be available in the `$PATH` which they aren't.  As we do not consume or ship these files from the build, add a try/catch block to the task so that it does not fail the build if it fails.